refactor: StockPriceDataVO 클래스 이름을 StockPriceManager로 변경
-  VO 객체는 단순한 값객체 취급용 , 도메인 맥락을 표현하지 X
