+++ src/main/java/com/flab/mars/domain/service/StockService.java
@@ -7,7 +7,6 @@ import com.flab.mars.db.entity.PriceDataEntity;
 import com.flab.mars.db.entity.StockInfoEntity;
 import com.flab.mars.db.repository.PriceDataRepository;
 import com.flab.mars.db.repository.StockInfoRepository;
-import com.flab.mars.domain.vo.StockPriceDataVO;
 import com.flab.mars.domain.vo.TokenInfoVO;
 import com.flab.mars.domain.vo.response.PriceDataVO;
 import com.flab.mars.domain.vo.response.StockFluctuationVO;
@@ -29,7 +28,7 @@ public class StockService {
 
     private final StockInfoRepository stockInfoRepository;
 
-    private final StockPriceDataVO stockPriceDataVO;
+    private final StockPriceManager stockPriceManager;
 
 
     public PriceDataVO getStockPrice(String stockCode, TokenInfoVO tokenInfo) {
@@ -48,7 +47,7 @@ public class StockService {
 
         KisStockPriceDto stockPrice = kisClient.getStockPrice(tokenInfo.getAccessToken(), tokenInfo.getAppKey(), tokenInfo.getAppSecret(), stockCode);
 
-        return stockPriceDataVO.saveCurrentStockPrice(stockPrice, stockInfo, currentTime);
+        return stockPriceManager.saveCurrentStockPrice(stockPrice, stockInfo, currentTime);
     }
 
     public StockFluctuationVO getFluctuationRanking(String url, TokenInfoVO tokenInfo) {
